follow-up changes

+++ src/builders/QueryBuilder.js
@@ -37,14 +37,15 @@ class QueryBuilder {
     }
 
     getBy({ column, matchingValue, projections = [] }) {
-        if (!column && matchingValue) return;
-        console.log(`Fetching matched records from "${this.tableName}" table with column:${this.#convertSnakeToCamel(column)},value:${matchingValue}`);
-        this.query = this.query.select(projections.join(",")).eq(this.#convertSnakeToCamel(column), matchingValue);
+        if (!column || !matchingValue) return;
+        let newColumnName = this.#convertCamelToSnake(column);
+        console.log(`Fetching matched records from "${this.tableName}" table with column:${newColumnName},value:${matchingValue}`);
+        this.query = this.query.select(projections.join(",")).eq(newColumnName, matchingValue);
         return this;
     }
 
     update({ column, matchingValue, data }) {
-        if(!column && matchingValue) return;
+        if (!column || !matchingValue) return;
         console.log(`Updating a record using column:${this.#convertCamelToSnake(column)} matching with:${matchingValue}`);
         let serilizedData = this.#serilizeData(data);
         this.query = this.query.update(serilizedData).eq(this.#convertCamelToSnake(column), matchingValue);
@@ -52,7 +53,7 @@ class QueryBuilder {
     }
 
     delete({ column, matchingValue }) {
-        if (!column && matchingValue) return;
+        if (!column || !matchingValue) return;
         let newColumnName = this.#convertCamelToSnake(column);
         console.log(`Deleting a record using column:${newColumnName} matching with:${matchingValue}`);
         this.query = this.query.delete().eq(newColumnName, matchingValue);
@@ -60,7 +61,7 @@ class QueryBuilder {
     }
 
     insert(dataObject) {
-        if (!column && matchingValue) return;
+        if (!column || !matchingValue) return;
         console.log(`Inserting a record into "${this.tableName}" table with data:${JSON.stringify(dataObject)}`);
         let serilizedData = this.#serilizeData(dataObject);
         this.query = this.query.insert(serilizedData);
